backend/app/requirements.txt
werkzeug==2.0.3
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
boto3==1.26.137
Pillow==11.0.0
opencv-python==4.10.0.84